flutter_backebnd/flutter_application_1/requirements.txt
flask==3.0.3

flask-cors==4.0.0

google-api-python-client==2.123.0

google-auth==2.29.0

requests==2.32.3

python-dotenv==1.0.1

werkzeug==3.0.3  # Specific compatible version

gspread==6.1.4